refactor: websys_bridge
